Ledger.Index

Description

An index of unspent transaction outputs, and some functions for validating transactions using the UTXO index.

Documentation

type ValidationMonad m = (MonadReader UtxoIndex m, MonadError ValidationError m)

Context for validating transactions. We need access to the unspent transaction outputs of the blockchain, and we can throw ValidationErrors

newtype UtxoIndex

The transactions of a blockchain indexed by hash

Constructors

UtxoIndex

Fields

getIndex :: Map TxOutRef TxOut

empty :: UtxoIndex

An empty [[UtxoIndex]]

insert :: Tx -> UtxoIndex -> UtxoIndex

Add a transaction to the index

insertBlock :: [Tx] -> UtxoIndex -> UtxoIndex

Add a block of transactions to the index

initialise :: Blockchain -> UtxoIndex

Create an index of all transactions on the chain

data Validation a

runValidation :: Validation a -> UtxoIndex -> Either ValidationError a

Run a Validation on a UtxoIndex

lkpValue :: ValidationMonad m => TxOutRef -> m Value

Determine the unspent value that a [[TxOutRef]] refers to

lkpTxOut :: ValidationMonad m => TxOutRef -> m TxOut

Find an unspent transaction output by its reference. Assumes that the output for this reference exists. If you want to handle the lookup error you can use runLookup. Determine the transaction output that a [[TxOutRef']] refers to

lkpOutputs :: ValidationMonad m => Tx -> m [(TxIn, TxOut)]

Match each input of the transaction with its output

data ValidationError

Reason why a transaction is invalid

Constructors

InOutTypeMismatch TxIn TxOut
A pay-to-pubkey output was consumed by a pay-to-script input or vice versa

TxOutRefNotFound TxOutRef
The unspent transaction output consumed by a transaction input could not be found (either because it was already spent, or because there was no transaction with the given hash on the blockchain)

InvalidScriptHash DataScript
(for pay-to-script outputs) The validator script provided in the transaction input does not match the hash specified in the transaction output

InvalidSignature PubKey Signature
(for pay-to-pubkey outputs) The signature of the transaction input does not match the public key of the transaction output

ValueNotPreserved Value Value
The amount spent by the transaction differs from the amount consumed by it

NegativeValue Tx
The transaction produces an output with a negative value

ScriptFailure [String]
(for pay-to-script outputs) Evaluation of the validator script failed

data InOutMatch

Matching pair of transaction input and transaction output.

Constructors

ScriptMatch TxIn ValidatorScript RedeemerScript DataScript (AddressOf (Digest SHA256))

PubKeyMatch PubKey Signature

Actual validation

validateTransaction :: ValidationMonad m => Slot -> Tx -> m ()

Validate a transaction in a ValidationMonad context.

This does the same as validTx, but more efficiently as it doesn't compute the UTXO of a blockchain from scratch. It also gives a more precise error: ValidationError instead of False.