Ledger.Interval

Description

A type for intervals and associated functions.

Documentation

newtype Slot

Slot number

Constructors

Slot

Fields

getSlot :: Int

data Interval a

An interval of as. The interval is closed below and open above, meaning that Interval (Just (10 :: Int)) (Just 11) contains a single value 11. The interval is unbounded on either side: Interval Nothing (Just 12) contains all numbers smaller than 12.

Constructors

Interval

Fields

ivFrom :: Maybe a

ivTo :: Maybe a

type SlotRange = Interval Slot

interval :: Q (TExp (a -> a -> Interval a))

interval a b includes all values that are greater than or equal to a and smaller than b. Therefore it includes a but not it does not include b.

from :: Q (TExp (a -> Interval a))

from a is an Interval that includes all values that are greater than or equal to a.

to :: Q (TExp (a -> Interval a))

to a is an Interval that includes all values that are smaller than a.

singleton :: Q (TExp (Slot -> SlotRange))

A SlotRange that covers only a single slot.

empty :: Q (TExp (SlotRange -> Bool))

Check if a SlotRange is empty

always :: Q (TExp SlotRange)

A SlotRange that covers every slot.

member :: Q (TExp (Slot -> SlotRange -> Bool))

Check if Slot is contained in a SlotRange.

width :: Q (TExp (SlotRange -> Maybe Int))

Number of Slots covered by the interval. width (from x) == Nothing

contains :: Q (TExp (SlotRange -> SlotRange -> Bool))

a contains b is true if the SlotRange b is entirely contained in a. That is, a contains b if for every slot s, if member s b then member s a.

before :: Q (TExp (Slot -> SlotRange -> Bool))

Check if a Slot is earlier than the beginning of a SlotRange

after :: Q (TExp (Slot -> SlotRange -> Bool))

Check if a Slot is later than the end of a SlotRange