Ledger.Types

Basic types

newtype Value

Cryptocurrency value

Constructors

Value

Fields

getValue :: Int

newtype Slot

Slot number

Constructors

Slot

Fields

getSlot :: Int

lastSlot :: Blockchain -> Slot

The number of the last slot of a blockchain. Assumes that empty slots are represented as empty blocks (as opposed to no blocks). This is true in the emulator but not necessarily on the real chain,

newtype TxIdOf h

Transaction ID (double SHA256 hash of the transaction)

Constructors

TxIdOf

Fields

getTxId :: h

type TxId = TxIdOf (Digest SHA256)

newtype PubKey

Public key

Constructors

PubKey

Fields

getPubKey :: Int

newtype Signature

Constructors

Signature

Fields

getSignature :: Int

signedBy :: Signature -> PubKey -> Bool

True if the signature matches the public key

Addresses

newtype AddressOf h

A payment address is a double SHA256 of a UTxO output's validator script (and presumably its data script). This corresponds to a Bitcoing pay-to-witness-script-hash

Constructors

AddressOf

Fields

getAddress :: h

type Address = AddressOf (Digest SHA256)

pubKeyAddress :: PubKey -> AddressOf (Digest SHA256)

The address of a transaction output locked by public key

scriptAddress :: ValidatorScript -> AddressOf (Digest SHA256)

The address of a transaction output locked by a validator script

Scripts

data Script

Script

fromCompiledCode :: CompiledCode a -> Script

lifted :: Lift a => a -> Script

applyScript :: Script -> Script -> Script

evaluateScript :: Script -> ([String], Bool)

newtype ValidatorScript

A validator is a PLC script.

Constructors

ValidatorScript

Fields

getValidator :: Script

newtype RedeemerScript

Redeemer (supplied by consumer of the transaction output)

Constructors

RedeemerScript

Fields

getRedeemer :: Script

newtype DataScript

Data script (supplied by producer of the transaction output)

Constructors

DataScript

Fields

getDataScript :: Script

Transactions

data Tx

Transaction including witnesses for its inputs

Constructors

Tx

Fields

txInputs :: Set TxIn

txOutputs :: [TxOut]

txForge :: !Value

txFee :: !Value

data TxStripped

Transaction without witnesses for its inputs

Constructors

TxStripped

Fields

txStrippedInputs :: Set TxOutRef

txStrippedOutputs :: [TxOut]

txStrippedForge :: !Value

txStrippedFee :: !Value

strip :: Tx -> TxStripped

preHash :: TxStripped -> Digest SHA256

Hash a stripped transaction once

hashTx :: Tx -> TxId

Double hash of a transaction, excluding its witnesses

dataTxo :: Blockchain -> TxOutRef -> Maybe DataScript

Determine the data script that an input refers to

data TxInOf h

Transaction input

Constructors

TxInOf

Fields

txInRef :: !(TxOutRefOf h)

txInType :: !TxInType

data TxInType

Type of transaction input.

Constructors

ConsumeScriptAddress !ValidatorScript !RedeemerScript

ConsumePublicKeyAddress !Signature

type TxIn = TxInOf (Digest SHA256)

data TxOutOf h

Constructors

TxOutOf

Fields

txOutAddress :: !(AddressOf h)

txOutValue :: !Value

txOutType :: !TxOutType

data TxOutType

Type of transaction output.

Constructors

PayToScript !DataScript
A pay-to-script output with the data script

PayToPubKey !PubKey
A pay-to-pubkey output

type TxOut = TxOutOf (Digest SHA256)

data TxOutRefOf h

Reference to a transaction output

Constructors

TxOutRefOf

Fields

txOutRefId :: TxIdOf h

txOutRefIdx :: Int

Index into the referenced transaction's outputs

type TxOutRef = TxOutRefOf (Digest SHA256)

simpleInput :: TxOutRefOf a -> TxInOf a

Transaction input that spends an output using the empty validator and unit redeemer scripts.

simpleOutput :: Value -> TxOut

Transaction output locked by the empty validator and unit data scripts.

pubKeyTxIn :: TxOutRefOf h -> Signature -> TxInOf h

scriptTxIn :: TxOutRefOf h -> ValidatorScript -> RedeemerScript -> TxInOf h

pubKeyTxOut :: Value -> PubKey -> TxOut

Create a transaction output locked by a public key

scriptTxOut :: Value -> ValidatorScript -> DataScript -> TxOut

Create a transaction output locked by a validator script

isPubKeyOut :: TxOutOf h -> Bool

Returns true if the output is a pay-to-pubkey output

isPayToScriptOut :: TxOutOf h -> Bool

Returns true if the output is a pay-to-script output

txOutRefs :: Tx -> [(TxOut, TxOutRef)]

A list of a transaction's outputs paired with their [[TxOutRef]]s

Blockchain & UTxO model

type Block = [Tx]

type Blockchain = [Block]

data BlockchainState

Ledger and transaction state available to both the validator and redeemer scripts

Constructors

BlockchainState

Fields

blockchainSlot :: Slot

blockchainStateTxHash :: TxId

newtype ValidationData

Information about the state of the blockchain and about the transaction that is currently being validated, represented as a value in PLC.

In the future we will generate this at transaction validation time, but for now we have to construct it at compilation time (in the test suite) and pass it to the validator.

Constructors

ValidationData Script

state :: Tx -> Blockchain -> BlockchainState

Get blockchain state for a transaction

transaction :: Blockchain -> TxOutRef -> Maybe Tx

Lookup a transaction by its hash

out :: Blockchain -> TxOutRef -> Maybe TxOut

Determine the unspent output that an input refers to

value :: Blockchain -> TxOutRef -> Maybe Value

Determine the unspent value that an input refers to

unspentOutputsTx :: Tx -> Map TxOutRef TxOut

The unspent outputs of a transaction

spentOutputs :: Tx -> Set TxOutRef

The outputs consumed by a transaction

unspentOutputs :: Blockchain -> Map TxOutRef TxOut

Unspent outputs of a ledger.

updateUtxo :: Tx -> Map TxOutRef TxOut -> Map TxOutRef TxOut

Update a map of unspent transaction outputs and sigantures with the inputs and outputs of a transaction.

validTx :: ValidationData -> Tx -> Blockchain -> Bool

Determine whether a transaction is valid in a given ledger

The inputs refer to unspent outputs, which they unlock (input validity).

The transaction preserves value (value preservation).

All values in the transaction are non-negative.

txOutPubKey :: TxOutOf h -> Maybe PubKey

The public key that a [[TxOut]] refers to

pubKeyTxo :: Blockchain -> TxOutRef -> Maybe PubKey

Determine the public key that locks the txo

validValuesTx :: Tx -> Bool

Check that all values in a transaction are no.

Scripts

validate :: ValidationData -> TxIn -> TxOut -> Bool

Check whether a transaction output can be spent by the given transaction input. This involves

Checking that pay-to-script (P2S) output is spent by a P2S input, and a pay-to-public key (P2PK) output is spend by a P2PK input

If it is a P2S input:

Verifying the hash of the validator script

Evaluating the validator script with the redeemer and data script

If it is a P2PK input:

Verifying that the signature matches the public key

emptyValidator :: ValidatorScript

() () () -> () as a validator

NB. The signature of a validator script is *d -> r -> v -> ()* where *d*, *r* and *v* are the (PLC) types of data script, redeemer script, and validation data. *d*, *r* and *v* are determined by the validator script itself (because applying it to values of different types will cause the script to be ill-typed). As a result, if you lock a transaction output with emptyValidator, you need to provide unitData, unitRedeemer and unitValidationData to consume it.

unitRedeemer :: RedeemerScript

() as a redeemer

unitData :: DataScript

() as a data script

unitValidationData :: ValidationData

() as validation data

runScript :: ValidationData -> ValidatorScript -> RedeemerScript -> DataScript -> ([String], Bool)

Evaluate a validator script with the given inputs

Lenses

inputs :: Lens' Tx (Set TxIn)

The inputs of a transaction

outputs :: Lens' Tx [TxOut]

outAddress :: Lens (TxOutOf h) (TxOutOf g) (AddressOf h) (AddressOf g)

The address of a transaction output

outValue :: Lens' (TxOutOf h) Value

The value of a transaction output | TODO: Compute address again

outType :: Lens' (TxOutOf h) TxOutType

The type of a transaction output | TODO: Compute address again

inRef :: Lens (TxInOf h) (TxInOf g) (TxOutRefOf h) (TxOutRefOf g)

The TxOutRefOf spent by a transaction input

inType :: Lens' (TxInOf h) TxInType

The type of a transaction input

inScripts :: TxInOf h -> Maybe (ValidatorScript, RedeemerScript)

Validator and redeemer scripts of a transaction input that spends a "pay to script" output

inSignature :: TxInOf h -> Maybe Signature

Signature of a transaction input that spends a "pay to public key" output

Orphan instances

Serialise (Digest SHA256)

Instance details

Methods

encode :: Digest SHA256 -> Encoding

decode :: Decoder s (Digest SHA256)

encodeList :: [Digest SHA256] -> Encoding

decodeList :: Decoder s [Digest SHA256]

FromJSON (Digest SHA256)

Instance details

Methods

parseJSON :: Value -> Parser (Digest SHA256)

parseJSONList :: Value -> Parser [Digest SHA256]

ToJSON (Digest SHA256)

Instance details

Methods

toJSON :: Digest SHA256 -> Value

toEncoding :: Digest SHA256 -> Encoding

toJSONList :: [Digest SHA256] -> Value

toEncodingList :: [Digest SHA256] -> Encoding

ToSchema (Digest SHA256)

Instance details

Methods

declareNamedSchema :: proxy (Digest SHA256) -> Declare (Definitions Schema) NamedSchema