Ledger.Validation

Pending transactions and related types

data PendingTx

A pending transaction as seen by validator scripts.

Constructors

PendingTx

Fields

pendingTxInputs :: [PendingTxIn]

Transaction inputs

pendingTxOutputs :: [PendingTxOut]

pendingTxFee :: Value

pendingTxForge :: Value

pendingTxSlot :: Slot

pendingTxIn :: PendingTxIn

PendingTxIn being validated

data PendingTxOut

Output of a pending transaction.

Constructors

PendingTxOut

Fields

pendingTxOutValue :: Value

pendingTxOutHashes :: Maybe (ValidatorHash, DataScriptHash)

Hashes of validator script and data script

pendingTxOutData :: PendingTxOutType

data PendingTxOutRef

Constructors

PendingTxOutRef

Fields

pendingTxOutRefId :: TxHash

Transaction whose outputs are consumed

pendingTxOutRefIdx :: Int

Index into the referenced transaction's list of outputs

data PendingTxIn

Input of a pending transaction.

Constructors

PendingTxIn

Fields

pendingTxInRef :: PendingTxOutRef

pendingTxInWitness :: Either (ValidatorHash, RedeemerHash) Signature

Tx input witness, hashes for Script input, or signature for a PubKey

pendingTxInValue :: Value

Value consumed by this txn input

data PendingTxOutType

Constructors

PubKeyTxOut PubKey
Pub key address

DataTxOut
The data script of the pending transaction output (see note [Script types in pending transactions])

Hashes (see note [Hashes in validator scripts])

newtype DataScriptHash

Constructors

DataScriptHash ByteString

newtype RedeemerHash

Constructors

RedeemerHash ByteString

newtype ValidatorHash

PLC runtime representation of a `Digest SHA256`

Constructors

ValidatorHash ByteString

newtype TxHash

Constructors

TxHash ByteString

plcDataScriptHash :: DataScript -> DataScriptHash

plcValidatorDigest :: Digest SHA256 -> ValidatorHash

plcRedeemerHash :: RedeemerScript -> RedeemerHash

plcTxHash :: TxId -> TxHash

Oracles

data OracleValue a

Constructors

OracleValue

Fields

ovSignature :: PubKey

ovSlot :: Slot

ovValue :: a

Validator functions

Signatures

txSignedBy :: Q (TExp (PendingTx -> PubKey -> Bool))

Check if a transaction was signed by a public key

txInSignedBy :: Q (TExp (PendingTxIn -> PubKey -> Bool))

Check if the input of a pending transaction was signed by a public key

Transactions

pubKeyOutput :: Q (TExp (PendingTxOut -> Maybe PubKey))

Returns the public key that locks the transaction output

scriptOutput :: Q (TExp (PendingTxOut -> Maybe (ValidatorHash, DataScriptHash)))

Returns the data script that is part of the pay-to-script transaction output

eqPubKey :: Q (TExp (PubKey -> PubKey -> Bool))

Equality of public keys

eqDataScript :: Q (TExp (DataScriptHash -> DataScriptHash -> Bool))

Equality of data scripts

eqRedeemer :: Q (TExp (RedeemerHash -> RedeemerHash -> Bool))

Equality of redeemer scripts

eqValidator :: Q (TExp (ValidatorHash -> ValidatorHash -> Bool))

Equality of validator scripts

eqTx :: Q (TExp (TxHash -> TxHash -> Bool))

Equality of transactions

Hashes

plcSHA2_256 :: ByteString -> ByteString

PLC-compatible SHA-256 hash of a hashable value

plcSHA3_256 :: ByteString -> ByteString

PLC-compatible SHA3-256 hash of a hashable value