Wallet.API

Description

Interface between wallet and Plutus client

Documentation

class WalletAPI m where

Used by Plutus client to interact with wallet

Minimal complete definition

submitTxn, myKeyPair, createPaymentWithChange, register, watchedAddresses, startWatching, slot

Methods

submitTxn :: Tx -> m ()

myKeyPair :: m KeyPair

createPaymentWithChange :: Value -> m (Set TxIn, Maybe TxOut)

Create a payment that spends the specified value and returns any leftover funds as change. Fails if we don't have enough funds.

register :: EventTrigger -> EventHandler m -> m ()

Register a [[EventHandler]] in `m ()` to be run when condition is true.

The action will be run once for each block where the condition holds. For example, `register (slotRangeT (Interval 3 6)) a` causes a to be run at blocks 3, 4, and 5.

Each time the wallet is notified of a new block, all triggers are checked and the matching ones are run in an unspecified order.

The wallet will only watch "known" addresses. There are two ways an address can become a known address. 1. When a trigger is registered for it 2. When a transaction submitted by this wallet produces an output for it When an address is added to the set of known addresses, it starts out with an initial value of 0. If there already exist unspent transaction outputs at that address on the chain, then those will be ignored.

`register c a >> register c b = register c (a >> b)`

watchedAddresses :: m AddressMap

startWatching :: Address -> m ()

slot :: m Slot

class MonadError WalletAPIError m => WalletDiagnostics m where

The ability to log messages and throw errors

Minimal complete definition

logMsg

Methods

logMsg :: Text -> m ()

Write some information to the log

newtype EventHandler m

An action than can be run in response to a blockchain event. It receives its [[EventTrigger]] annotated with truth values.

Constructors

EventHandler

Fields

runEventHandler :: AnnotatedEventTrigger Bool -> m ()

newtype KeyPair

Constructors

KeyPair

Fields

getKeyPair :: (PrivateKey, PubKey)

newtype PubKey

Public key

Constructors

PubKey

Fields

getPubKey :: Int

pubKey :: KeyPair -> PubKey

Get the public key of a [[KeyPair]]

keyPair :: Int -> KeyPair

Create a [[KeyPair]] given a "private key"

signature :: KeyPair -> Signature

Create a [[Signature]] signed by the private key of a [[KeyPair]]

createPayment :: (Functor m, WalletAPI m) => Value -> m (Set TxIn)

createTxAndSubmit :: (Monad m, WalletAPI m) => SlotRange -> Set TxIn -> [TxOut] -> m Tx

Create a transaction and submit it. TODO: Also compute the fee

payToScript :: (Monad m, WalletAPI m) => SlotRange -> Address -> Value -> DataScript -> m Tx

Transfer some funds to an address locked by a script, returning the transaction that was submitted.

payToScript_ :: (Monad m, WalletAPI m) => SlotRange -> Address -> Value -> DataScript -> m ()

Transfer some funds to an address locked by a script.

payToPublicKey :: (Monad m, WalletAPI m) => SlotRange -> Value -> PubKey -> m Tx

Transfer some funds to an address locked by a public key, returning the transaction that was submitted.

payToPublicKey_ :: (Monad m, WalletAPI m) => SlotRange -> Value -> PubKey -> m ()

Transfer some funds to an address locked by a public key

payToScripts :: (Monad m, WalletAPI m) => SlotRange -> [(Address, Value, DataScript)] -> m Tx

Transfer some funds to a number of script addresses, returning the transaction that was submitted.

payToScripts_ :: (Monad m, WalletAPI m) => SlotRange -> [(Address, Value, DataScript)] -> m ()

Transfer some funds to a number of script addresses.

collectFromScript :: (Monad m, WalletAPI m) => SlotRange -> ValidatorScript -> RedeemerScript -> m ()

Collect all unspent outputs from a pay to script address and transfer them to a public key owned by us.

collectFromScriptTxn :: (Monad m, WalletAPI m) => SlotRange -> ValidatorScript -> RedeemerScript -> TxId -> m ()

Given the pay to script address of the ValidatorScript, collect from it all the inputs that were produced by a specific transaction, using the RedeemerScript.

ownPubKeyTxOut :: (Monad m, WalletAPI m) => Value -> m TxOut

Create a TxOut that pays to a public key owned by us

ownPubKey :: (Functor m, WalletAPI m) => m PubKey

Get the public key for this wallet

ownSignature :: (Functor m, WalletAPI m) => m Signature

Generate a Signature with the wallet's own private key

Slot ranges

data Interval a

An interval of as. The interval is closed below and open above, meaning that Interval (Just (10 :: Int)) (Just 11) contains a single value 11. The interval is unbounded on either side: Interval Nothing (Just 12) contains all numbers smaller than 12.

Constructors

Interval

Fields

ivFrom :: Maybe a

ivTo :: Maybe a

type SlotRange = Interval Slot

defaultSlotRange :: SlotRange

An open-ended SlotRange that begins at slot 1.

interval :: a -> a -> Interval a

An inclusive-exclusive interval

intervalFrom :: a -> Interval a

intervalFrom a includes all values greater than or equal to a, including a itself.

intervalTo :: a -> Interval a

intervalTo a includes all values smaller than a, but not a itself.

singleton :: Slot -> SlotRange

A SlotRange that covers a single slot

empty :: SlotRange -> Bool

Check whether a SlotRange is empty

always :: SlotRange

A SlotRange that covers all the slots

member :: Ord a => a -> Interval a -> Bool

Check whether an Interval a includes an a.

width :: SlotRange -> Maybe Int

The number of slots included in the SlotRange

before :: Slot -> SlotRange -> Bool

Check if a Slot is before a SlotRange

after :: Slot -> SlotRange -> Bool

Check if a Slot is after a SlotRange

contains :: SlotRange -> SlotRange -> Bool

Triggers

type EventTrigger = Fix EventTriggerF

type AnnotatedEventTrigger a = Fix (Compose ((,) a) EventTriggerF)

An [[EventTrigger]] where each level is annotated with a value of a

data EventTriggerF f

Constructors

TAnd f f

TOr f f

TNot f

TAlways

TNever

TSlotRange !SlotRange

TFundsAtAddress !Address !(Interval Value)

andT :: EventTrigger -> EventTrigger -> EventTrigger

`andT l r` is true when l and r are true.

orT :: EventTrigger -> EventTrigger -> EventTrigger

`orT l r` is true when l or r are true.

notT :: EventTrigger -> EventTrigger

`notT t` is true when t is false

alwaysT :: EventTrigger

alwaysT is always true

neverT :: EventTrigger

neverT is never true

slotRangeT :: SlotRange -> EventTrigger

`slotRangeT r` is true when the slot number is in the range r

fundsAtAddressT :: Address -> Interval Value -> EventTrigger

`fundsAtAddressT a r` is true when the funds at a are in the range r

checkTrigger :: Slot -> Map Address Value -> EventTrigger -> Bool

Check if the given slot number and UTXOs match the conditions of an [[EventTrigger]]

addresses :: EventTrigger -> [Address]

The addresses that an [[EventTrigger]] refers to

getAnnot :: AnnotatedEventTrigger a -> a

annTruthValue :: Slot -> Map Address Value -> EventTrigger -> AnnotatedEventTrigger Bool

Annotate each node in an EventTriggerF with its truth value given a slot and a set of unspent outputs

Error handling

data WalletAPIError

Constructors

InsufficientFunds Text

OtherError Text

throwInsufficientFundsError :: MonadError WalletAPIError m => Text -> m a

throwOtherError :: MonadError WalletAPIError m => Text -> m a

Logging

newtype WalletLog

Constructors

WalletLog

Fields

getWalletLog :: [Text]