Wallet.Emulator.AddressMap

Documentation

newtype AddressMap

A map of [[Address]]es and their unspent outputs

Constructors

AddressMap

Fields

getAddressMap :: Map Address (Map TxOutRef TxOut)

addAddress :: Address -> AddressMap -> AddressMap

Add an address with no unspent outputs. If the address already exists, do nothing.

addAddresses :: [Address] -> AddressMap -> AddressMap

Add a list of [[Address]]es with no unspent outputs

values :: AddressMap -> Map Address Value

The total value of unspent outputs at an address

fromTxOutputs :: Tx -> AddressMap

An [[AddressMap]] with the unspent outputs of a single transaction

knownAddresses :: AddressMap -> Map TxOutRef Address

A map of unspent transaction outputs to their addresses (the "inverse" of [[AddressMap]], without the values

updateAddresses :: Tx -> AddressMap -> AddressMap

Update the [[AddressMap]] with the inputs and outputs of a new transaction. updateAddresses does not add or remove any keys from its second argument.

restrict :: AddressMap -> Set Address -> AddressMap

Restrict an [[AddressMap]] to a set of addresses.