Wallet.Emulator.Types

Wallets

newtype Wallet

Constructors

Wallet

Fields

getWallet :: Int

type TxPool = [Tx]

Emulator

data Assertion

Constructors

IsValidated Tx

OwnFundsEqual Wallet Value

assert :: MonadEmulator m => Assertion -> m ()

assertIsValidated :: Tx -> Trace m ()

data AssertionError

data Event n a where

The type of events in the emulator. n is the type (usually a monad) in which wallet actions take place.

Constructors

WalletAction :: Wallet -> n () -> Event n [Tx]
An direct action performed by a wallet. Usually represents a "user action", as it is triggered externally.

WalletRecvNotification :: Wallet -> [Notification] -> Event n [Tx]
A wallet receiving some notifications, and reacting to them.

BlockchainProcessPending :: Event n Block
The blockchain processing pending transactions, producing a new block from the valid ones and discarding the invalid ones.

Assertion :: Assertion -> Event n ()
An assertion in the event stream, which can inspect the current state.

data Notification

Constructors

BlockValidated Block

CurrentSlot Slot

data EmulatorEvent

Events produced by the mockchain

Constructors

TxnSubmit TxId
A transaction has been added to the global pool of pending transactions

TxnValidate TxId
A transaction has been validated and added to the blockchain

TxnValidationFail TxId ValidationError
A transaction failed to validate

SlotAdd Slot
A slot has passed, and a block was added to the blokchain

WalletError Wallet WalletAPIError
A WalletAPI action produced an error

WalletInfo Wallet Text
Debug information produced by a wallet

Wallet state

data WalletState

Constructors

WalletState

Fields

_ownKeyPair :: KeyPair

_walletSlot :: Slot

Current slot as far as the wallet is concerned

_addressMap :: AddressMap

Addresses that we watch. For each address we keep the unspent transaction outputs and their values, so that we can use them in transactions.

_triggers :: Map EventTrigger (EventHandler MockWallet)

emptyWalletState :: Wallet -> WalletState

An empty wallet state with the public/private key pair for a wallet, and the public key address for that wallet as the sole member of walletStateWatchedAddresses

ownKeyPair :: Lens' WalletState KeyPair

ownFunds :: Lens' WalletState (Map TxOutRef TxOut)

addressMap :: Lens' WalletState AddressMap

walletSlot :: Lens' WalletState Slot

Traces

type Trace m = Program (Event m)

runTraceChain :: Blockchain -> Trace MockWallet a -> (Either AssertionError a, EmulatorState)

Run an emulator trace on a blockchain

runTraceTxPool :: TxPool -> Trace MockWallet a -> (Either AssertionError a, EmulatorState)

Run an emulator trace on an empty blockchain with a pool of pending transactions

walletAction :: Wallet -> m () -> Trace m [Tx]

Interact with a wallet

walletRecvNotifications :: Wallet -> [Notification] -> Trace m [Tx]

Notify a wallet of blockchain events

walletNotifyBlock :: Wallet -> Block -> Trace m [Tx]

Notify a wallet that a block has been validated

walletsNotifyBlock :: [Wallet] -> Block -> Trace m [Tx]

Notify a list of wallets that a block has been validated

processPending :: Trace m Block

Validate all pending transactions

addBlocks :: Int -> Trace m [Block]

Add a number of empty blocks to the blockchain, by performing processPending n times.

addBlocksAndNotify :: [Wallet] -> Int -> Trace m ()

Add a number of blocks, notifying all wallets after each block

assertion :: Assertion -> Trace m ()

Make an assertion about the emulator state

assertOwnFundsEq :: Wallet -> Value -> Trace m ()

Emulator internals

newtype MockWallet a

Constructors

MockWallet

Fields

runMockWallet :: ExceptT WalletAPIError (StateT WalletState (Writer (WalletLog, [Tx]))) a

handleNotifications :: [Notification] -> MockWallet ()

data EmulatorState

Constructors

EmulatorState

Fields

_chainNewestFirst :: Blockchain

_txPool :: TxPool

_walletStates :: Map Wallet WalletState

_index :: UtxoIndex

_emulatorLog :: [EmulatorEvent]

emulator events, newest first

emptyEmulatorState :: EmulatorState

emulatorState :: Blockchain -> EmulatorState

Initialise the emulator state with a blockchain

chainNewestFirst :: Lens' EmulatorState Blockchain

chainOldestFirst :: Lens' EmulatorState Blockchain

The blockchain as a list of blocks, starting with the oldest (genesis) block

txPool :: Lens' EmulatorState TxPool

walletStates :: Lens' EmulatorState (Map Wallet WalletState)

index :: Lens' EmulatorState UtxoIndex

type MonadEmulator m = (MonadState EmulatorState m, MonadError AssertionError m)

validateEm :: EmulatorState -> Tx -> Maybe ValidationError

Validate a transaction in the current emulator state

validateBlock :: EmulatorState -> [Tx] -> ValidatedBlock

Validate a block in an [[EmulatorState]], returning the valid transactions and all success/failure events

data ValidatedBlock

Result of validating a block

Constructors

ValidatedBlock

Fields

vlbValid :: [Tx]

Transactions that have been validated

vlbEvents :: [EmulatorEvent]

Transaction validation events

vlbRest :: [Tx]

Transactions that haven't been validated because the current slot is not in their validation interval

liftMockWallet :: MonadState EmulatorState m => Wallet -> MockWallet a -> m ([Tx], Either WalletAPIError a)

evalEmulated :: MonadEmulator m => Event MockWallet a -> m a

processEmulated :: MonadEmulator m => Trace MockWallet a -> m a

runWalletActionAndProcessPending :: [Wallet] -> Wallet -> m () -> Trace m [Tx]

fundsDistribution :: EmulatorState -> Map Wallet Value

selectCoin :: MonadError WalletAPIError m => [(a, Value)] -> Value -> m ([(a, Value)], Value)

Given a set of as with coin values, and a target value, select a number of a such that their total value is greater than or equal to the target.