Wallet.Generators

Mockchain

data Mockchain

Blockchain for testing the emulator implementation and traces.

To avoid having to rely on functions from the implementation of wallet-api (in particular, unspentOutputs) we note the unspent outputs of the chain when it is first created.

Constructors

Mockchain

Fields

mockchainInitialBlock :: Block

mockchainUtxo :: Map TxOutRef TxOut

genMockchain :: MonadGen m => m Mockchain

Generate a mockchain using the default [[GeneratorModel]]

genMockchain' :: MonadGen m => GeneratorModel -> m Mockchain

Generate a mockchain

TODO: Generate more than 1 txn

emptyChain :: Mockchain

The empty mockchain

data GeneratorModel

Constructors

GeneratorModel

Fields

gmInitialBalance :: Map PubKey Value

Value created at the beginning of the blockchain

gmPubKeys :: Set PubKey

Public keys that are to be used for generating transactions

generatorModel :: GeneratorModel

A generator model with some sensible defaults

Transactions

newtype FeeEstimator

Estimate a transaction fee based on the number of its inputs and outputs.

Constructors

FeeEstimator

Fields

estimateFee :: Int -> Int -> Value

constantFee :: Value -> FeeEstimator

A constant fee for all transactions

genValidTransaction :: MonadGen m => Mockchain -> m Tx

Generate a valid transaction, using the unspent outputs provided. Fails if the there are no unspent outputs, or if the total value of the unspent outputs is smaller than the minimum fee (1)

genValidTransaction' :: MonadGen m => GeneratorModel -> FeeEstimator -> Mockchain -> m Tx

Generate a valid transaction, using the unspent outputs provided. Fails if the there are no unspent outputs, or if the total value of the unspent outputs is smaller than the estimated fee.

genValidTransactionSpending :: MonadGen m => Set TxIn -> Value -> m Tx

genValidTransactionSpending' :: MonadGen m => GeneratorModel -> FeeEstimator -> Set TxIn -> Value -> m Tx

genInitialTransaction :: GeneratorModel -> (Tx, [TxOut])

A transaction with no inputs that forges some value (to be used at the beginning of a blockchain)

Assertions

assertValid :: (MonadTest m, HasCallStack) => Tx -> Mockchain -> m ()

Assert that a transaction is valid in a chain

Etc.

genValue :: MonadGen m => m Value

runTrace :: Mockchain -> Trace MockWallet a -> (Either AssertionError a, EmulatorState)

Run an emulator trace on a mockchain

runTraceOn :: MonadGen m => GeneratorModel -> Trace MockWallet a -> m (Either AssertionError a, EmulatorState)

Run an emulator trace on a mockchain generated by the model

splitVal :: (MonadGen m, Integral n) => Int -> n -> m [n]

Split a value into max. n positive-valued parts such that the sum of the parts equals the original value.

validateMockchain :: Mockchain -> Tx -> Maybe ValidationError

Validate a transaction in a mockchain